Split ScrollableBalances.tsx into three files
